requirements.txt
fastapi[all]
passlib[bcrypt]
python-jose[cryptography]
pymongo
wsproto